requirements.txt
fastapi
uvicorn
openai
python-dotenv
pandas
openpyxl